requirements.txt
angelcommunity==2.0.0
bimlpa==0.1.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.3
dask==2023.9.3
decorator==5.1.1
demon==2.0.6
dynetx==0.3.2
eva-lcd==0.1.1
fonttools==4.43.1
fsspec==2023.9.2
future==0.18.3
idna==3.4
igraph==0.10.8
importlib-metadata==6.8.0
joblib==1.3.2
kiwisolver==1.4.5
leidenalg==0.10.1
Levenshtein==0.23.0
locket==1.0.0
markov-clustering==0.0.6.dev0
matplotlib==3.8.0
networkx==3.1
nf1==0.0.4
numpy==1.26.0
packaging==23.2
pandas==2.1.1
partd==1.4.1
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
PuLP==2.7.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-igraph==0.10.8
python-Levenshtein==0.23.0
python-louvain==0.16
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
setuptools==68.2.2
six==1.16.0
texttable==1.7.0
threadpoolctl==3.2.0
thresholdclustering==1.1
toolz==0.12.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.6
zipp==3.17.0
